Update script.js
In this version of the code, we used let and const instead of var, and we utilized the insertAdjacentHTML() function instead of creating a new element and appending it to the end of the list with appendChild(). We also placed some functions at the beginning of the code to make them easier to find. Finally, we added an event listener for the form to add new transactions.